Updated The General Test Scene with crawl space
I noticed that the platform prefab was broken so i created a new one with the same settings along with a wall prefab for simulating a crawlspace which is basicly a large square. I created the crawlspace attached it to the right of the walking test section where Austin made it above Adams level.